requirements.txt
# AI Resume Analyzer - ML Training Dependencies
# Python 3.10+ recommended

# Core ML Libraries
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Visualization & Analysis
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Jupyter & Development
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# SpaCy Language Model (download separately)
# python -m spacy download en_core_web_sm
google-generativeai>=0.3.0
openai>=1.0.0
SQLAlchemy>=2.0.0
